+++ src/collective/eeafaceted/z3ctable/browser/views.py
@@ -141,7 +141,7 @@ class FacetedTableView(BrowserView, ExtendedCSSTable):
         self.update(batch)
         try:
             return self.render()
-        except Exception, exc:
+        except Exception as exc:
             # in case an error occured, catch it or it freezes the web page
             # because faceted JS disable page and error raised does not unlock
             traceback.print_exc(None)