+++ src/collective/eeafaceted/z3ctable/browser/views.py
@@ -11,7 +11,7 @@ from z3c.table.interfaces import INoneCell
 from z3c.table.table import SequenceTable
 from zope.component import getAdapters
 from zope.component import queryMultiAdapter
-from zope.interface import implements
+from zope.interface import implementer
 
 import html
 import logging
@@ -112,15 +112,14 @@ class ExtendedCSSTable(SequenceTable):
         return renderedCell
 
 
+@implementer(IFacetedTable)
 class FacetedTableView(BrowserView, ExtendedCSSTable):
-
-    implements(IFacetedTable)
 
     # workaround so z3c.table does not manage batching
     startBatchingAt = 9999
     cssClassEven = u'odd'
     cssClassOdd = u'even'
-    cssClasses = {'table': 'faceted-table-results listing nosort'}
+    cssClasses = {'table': 'faceted-table-results table table-striped table-bordered nosort'}
     ignoreColumnWeight = False  # when set to True, keep columns ordered as returned by '_getViewFields'
     table_id = 'faceted_table'
     row_id_prefix = 'row_'