+++ CHANGES.rst
@@ -26,6 +26,8 @@ Changelog
 - In `DateColumn.renderCell`, no need to manage a `DateTime` as it is supported
   by `api.portal.get_localized_time`.
   [gbastien]
+- Fix Py3 Errors
+  [fngaha]
 
 2.24 (2023-10-19)
 -----------------